+++ backend/controllers/genre.controller.js
@@ -1,10 +1,18 @@
 const baseResponse = require("../dto/baseResponse");
 const {genreService} = require('../services/')
 const {StatusCodes} = require("http-status-codes");
+const helpers = require('../utils/helpers');
 
 exports.createGenre = async (req, res) => {
  try {
-  //invalida data here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.createGenre(req, res);
   res.status(StatusCodes.CREATED).json({
    ...baseResponse,
@@ -15,7 +23,7 @@ exports.createGenre = async (req, res) => {
   });
 
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in createGenre');
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -30,6 +38,14 @@ exports.createGenre = async (req, res) => {
 
 exports.getAllGenres = async (req, res) => {
  try {
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.getAllGenres();
   console.log(json)
   res.status(StatusCodes.OK).json({
@@ -41,7 +57,7 @@ exports.getAllGenres = async (req, res) => {
 
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in getAllGenres');
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -56,7 +72,14 @@ exports.getAllGenres = async (req, res) => {
 
 exports.getAllGenresWithPagination = async (req, res) => {
  try {
-  //invalid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.getAllGenresWithPagination(req, res);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -66,7 +89,7 @@ exports.getAllGenresWithPagination = async (req, res) => {
    code: StatusCodes.OK
   })
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in getAllGenresWithPagination')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error:true,
@@ -81,7 +104,14 @@ exports.getAllGenresWithPagination = async (req, res) => {
 
 exports.getById = async (req,res) => {
  try {
- // valid control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.getById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -91,7 +121,7 @@ exports.getById = async (req,res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in getById');
  res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
   ...baseResponse,
   error: true,
@@ -106,7 +136,14 @@ exports.getById = async (req,res) => {
 
 exports.updateById = async (req, res) => {
  try {
- //valid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.updateById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -116,7 +153,7 @@ exports.updateById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
- //log to error
+ helpers.logToError(error,req,'Error in updateById');
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -131,7 +168,14 @@ exports.updateById = async (req, res) => {
 
 exports.deleteById = async (req, res) => {
  try {
-  //invalid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await genreService.deleteById(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -141,7 +185,7 @@ exports.deleteById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error,req,'Error in deleteById')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
