+++ backend/controllers/user.controller.js
@@ -1,10 +1,18 @@
 const baseResponse = require("../dto/baseResponse");
 const userService = require("../services/user.service");
 const {StatusCodes} = require("http-status-codes");
+const helpers = require('../utils/helpers')
 
 exports.createUser = async (req, res) => {
  try {
-  //invalid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.createUser(req, res);
   res.status(StatusCodes.CREATED).json({
    ...baseResponse,
@@ -15,7 +23,7 @@ exports.createUser = async (req, res) => {
   });
  } catch (error) {
   console.log(error);
-  //log to error
+  helpers.logToError(error, req, 'Error in Create User')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -30,7 +38,14 @@ exports.createUser = async (req, res) => {
 
 exports.getAllUsers = async (req, res) => {
  try {
-  //invalid control here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.getAllUsers();
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -40,7 +55,7 @@ exports.getAllUsers = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  // log to error
+  helpers.logToError(error, req, 'Error in Get All Users')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -55,7 +70,14 @@ exports.getAllUsers = async (req, res) => {
 
 exports.getAllUsersWithPagination = async (req, res) => {
  try {
-// valida data control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.getAllUSersWithPagination(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -66,7 +88,7 @@ exports.getAllUsersWithPagination = async (req, res) => {
    code: StatusCodes.OK
   })
  } catch (error) {
-//log to error
+  helpers.logToError(error, req, 'Error in Get All Users With Pagination')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    success: false,
@@ -81,7 +103,14 @@ exports.getAllUsersWithPagination = async (req, res) => {
 
 exports.getUserById = async (req, res) => {
  try {
-  //valid control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.getById(req)
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -91,7 +120,7 @@ exports.getUserById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Get User By Id')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -106,7 +135,14 @@ exports.getUserById = async (req, res) => {
 
 exports.updateUserById = async (req, res) => {
  try {
-  //valid control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.updateUser(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -117,7 +153,7 @@ exports.updateUserById = async (req, res) => {
   });
 
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Update User By Id')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -132,7 +168,14 @@ exports.updateUserById = async (req, res) => {
 
 exports.deleteUserById = async (req, res) => {
  try {
-  //valid control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.deleteUser(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -142,7 +185,7 @@ exports.deleteUserById = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Delete User By Id')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -155,26 +198,33 @@ exports.deleteUserById = async (req, res) => {
 }
 
 
-exports.login = async (req,res) =>{
+exports.login = async (req, res) => {
  try {
- //valid control is here
-  const json = await userService.login(req,res);
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
+  const json = await userService.login(req, res);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
-   data:json,
-   success:true,
-   error:false,
+   data: json,
+   success: true,
+   error: false,
    timestamp: Date.now,
-   code:StatusCodes.OK
+   code: StatusCodes.OK
   })
  } catch (error) {
- //log to error
+  helpers.logToError(error, req, 'Error in Login')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
-   success:false,
-   error:true,
-   timestamp:Date.now(),
-   code:StatusCodes.INTERNAL_SERVER_ERROR,
+   success: false,
+   error: true,
+   timestamp: Date.now(),
+   code: StatusCodes.INTERNAL_SERVER_ERROR,
    message: error.message
   })
  }
@@ -183,7 +233,14 @@ exports.login = async (req,res) =>{
 
 exports.uploadPP = async (req, res) => {
  try {
-  //data
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.uploadProfilePicture(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -194,7 +251,7 @@ exports.uploadPP = async (req, res) => {
   });
 
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Upload Profile Picture')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -208,7 +265,14 @@ exports.uploadPP = async (req, res) => {
 
 exports.updatePP = async (req, res) => {
  try {
-  //validation error control is here
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.updateProfilePicture(req);
   res.status(StatusCodes.OK).json({
    ...baseResponse,
@@ -218,7 +282,7 @@ exports.updatePP = async (req, res) => {
    code: StatusCodes.OK
   });
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Update Profile Picture')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    error: true,
@@ -233,17 +297,25 @@ exports.updatePP = async (req, res) => {
 
 exports.addFriend = async (req, res) => {
  try {
+  const isInValid = helpers.handleValidationErrors(req);
+  if (isInValid) {
+   return res.status(StatusCodes.BAD_REQUEST).json({
+    ...baseResponse,
+    ...isInValid
+   })
+   return
+  }
   const json = await userService.addFriend(req, res);
- res.status(StatusCodes.OK).json({
-  ...baseResponse,
-  success:true,
-  data:json,
-  error:false,
-  timestamp: Date.now(),
-  code:StatusCodes.OK
- })
+  res.status(StatusCodes.OK).json({
+   ...baseResponse,
+   success: true,
+   data: json,
+   error: false,
+   timestamp: Date.now(),
+   code: StatusCodes.OK
+  })
  } catch (error) {
-  //log to error
+  helpers.logToError(error, req, 'Error in Add Friend')
   res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
    ...baseResponse,
    success: false,
